Mars_Project/mars-club project/metadata-generator/requirements.txt
transformers
torch
streamlit
pymupdf
groq
python-docx
python-dotenv